Teams are now league specific.
"Blue" can refer to both Cardiff Blues and the team from Super Rugby.
Move away from looking globally for team but consider the context that we're asking, i.e. the league.
This means that all fixture and table parsing now should be aware of the league that it is parsing.
Team is now found through the league rather than the data manager.